refactor run method
