+++ pkg/virtctl/imageupload/imageupload.go
@@ -87,77 +87,6 @@ var UploadProcessingCompleteFunc = waitUploadProcessingComplete
 // GetHTTPClientFn allows overriding the default http client (useful for unit testing)
 var GetHTTPClientFn = GetHTTPClient
 
-// NewImageUploadCommand returns a cobra.Command for handling the uploading of VM images
-func NewImageUploadCommand() *cobra.Command {
-	c := command{}
-	cmd := &cobra.Command{
-		Use:     "image-upload",
-		Short:   "Upload a VM image to a DataVolume/PersistentVolumeClaim.",
-		Example: usage(),
-		Args:    cobra.MaximumNArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			client, namespace, _, err := clientconfig.ClientAndNamespaceFromContext(cmd.Context())
-			if err != nil {
-				return err
-			}
-
-			c.cmd = cmd
-			c.client = client
-			c.namespace = namespace
-			return c.run(args)
-		},
-	}
-	cmd.Flags().BoolVar(&c.insecure, "insecure", false, "Allow insecure server connections when using HTTPS.")
-	cmd.Flags().StringVar(&c.uploadProxyURL, "uploadproxy-url", "", "The URL of the cdi-upload proxy service.")
-	cmd.Flags().StringVar(&c.name, "pvc-name", "", "The destination DataVolume/PVC name.")
-	cmd.Flags().StringVar(&c.pvcSize, "pvc-size", "", "The size of the PVC to create (ex. 10Gi, 500Mi).")
-	cmd.Flags().StringVar(&c.size, "size", "", "The size of the DataVolume to create (ex. 10Gi, 500Mi).")
-	cmd.Flags().StringVar(&c.storageClass, "storage-class", "", "The storage class for the PVC.")
-	cmd.Flags().StringVar(&c.accessMode, "access-mode", "", "The access mode for the PVC.")
-	cmd.Flags().BoolVar(&c.blockVolume, "block-volume", false, "Create a PVC with VolumeMode=Block (default is the storageProfile default. for archive upload default is filesystem).")
-	cmd.Flags().StringVar(&c.volumeMode, "volume-mode", "", "Specify the VolumeMode (block/filesystem) used to create the PVC. Default is the storageProfile default. For archive upload default is filesystem.")
-	cmd.Flags().StringVar(&c.imagePath, "image-path", "", "Path to the local VM image.")
-	cmd.Flags().StringVar(&c.archivePath, "archive-path", "", "Path to the local archive.")
-	cmd.Flags().BoolVar(&c.noCreate, "no-create", false, "Don't attempt to create a new DataVolume/PVC.")
-	cmd.Flags().UintVar(&c.uploadPodWaitSecs, "wait-secs", 300, "Seconds to wait for upload pod to start.")
-	cmd.Flags().UintVar(&c.uploadRetries, "retry", 5, "When upload server returns a transient error, we retry this number of times before giving up")
-	cmd.Flags().BoolVar(&c.forceBind, "force-bind", false, "Force bind the PVC, ignoring the WaitForFirstConsumer logic.")
-	cmd.Flags().BoolVar(&c.dataSource, "datasource", false, "Create a DataSource pointing to the created DataVolume/PVC.")
-	cmd.Flags().StringVar(&c.defaultInstancetype, "default-instancetype", "", "The default instance type to associate with the image.")
-	cmd.Flags().StringVar(&c.defaultInstancetypeKind, "default-instancetype-kind", "", "The default instance type kind to associate with the image.")
-	cmd.Flags().StringVar(&c.defaultPreference, "default-preference", "", "The default preference to associate with the image.")
-	cmd.Flags().StringVar(&c.defaultPreferenceKind, "default-preference-kind", "", "The default preference kind to associate with the image.")
-	cmd.SetUsageTemplate(templates.UsageTemplate())
-	cmd.Flags().MarkDeprecated("pvc-name", "specify the name as the second argument instead.")
-	cmd.Flags().MarkDeprecated("pvc-size", "use --size instead.")
-	cmd.Flags().MarkDeprecated("block-volume", "specify volume mode (filesystem/block) with --volume-mode instead.")
-	return cmd
-}
-
-func usage() string {
-	usage := `  # Upload a local disk image to a newly created DataVolume:
-  {{ProgramName}} image-upload dv fedora-dv --size=10Gi --image-path=/images/fedora30.qcow2
-
-  # Upload a local disk image to an existing DataVolume
-  {{ProgramName}} image-upload dv fedora-dv --no-create --image-path=/images/fedora30.qcow2
-
-  # Upload a local disk image to a newly created PersistentVolumeClaim
-  {{ProgramName}} image-upload pvc fedora-pvc --size=10Gi --image-path=/images/fedora30.qcow2
-
-  # Upload a local disk image to a newly created PersistentVolumeClaim and label it with a default instance type and preference
-  {{ProgramName}} image-upload pvc fedora-pvc --size=10Gi --image-path=/images/fedora30.qcow2 --default-instancetype=n1.medium --default-preference=fedora
-
-  # Upload a local disk image to an existing PersistentVolumeClaim
-  {{ProgramName}} image-upload pvc fedora-pvc --no-create --image-path=/images/fedora30.qcow2
-
-  # Upload to a DataVolume with explicit URL to CDI Upload Proxy
-  {{ProgramName}} image-upload dv fedora-dv --uploadproxy-url=https://cdi-uploadproxy.mycluster.com --image-path=/images/fedora30.qcow2
-
-  # Upload a local disk archive to a newly created DataVolume:
-  {{ProgramName}} image-upload dv fedora-dv --size=10Gi --archive-path=/images/fedora30.tar`
-	return usage
-}
-
 type command struct {
 	cmd                     *cobra.Command
 	client                  kubecli.KubevirtClient
@@ -184,126 +113,6 @@ type command struct {
 	forceBind               bool
 	dataSource              bool
 	archiveUpload           bool
-}
-
-func (c *command) validateDefaultInstancetypeArgs() error {
-	if c.defaultInstancetype == "" && c.defaultInstancetypeKind != "" {
-		return fmt.Errorf("--default-instancetype must be provided with --default-instancetype-kind")
-	}
-	if c.defaultPreference == "" && c.defaultPreferenceKind != "" {
-		return fmt.Errorf("--default-preference must be provided with --default-preference-kind")
-	}
-	if (c.defaultInstancetype != "" || c.defaultPreference != "") && c.noCreate {
-		return fmt.Errorf("--default-instancetype and --default-preference cannot be used with --no-create")
-	}
-	return nil
-}
-
-func (c *command) run(args []string) error {
-	if err := c.parseArgs(args); err != nil {
-		return err
-	}
-
-	if err := c.validateDefaultInstancetypeArgs(); err != nil {
-		return err
-	}
-
-	// #nosec G304 No risk for path injection as this function executes with
-	// the same privileges as those of virtctl user who supplies imagePath
-	file, err := os.Open(c.imagePath)
-	if err != nil {
-		return err
-	}
-	defer util.CloseIOAndCheckErr(file, nil)
-
-	pvc, err := c.getAndValidateUploadPVC()
-	if err != nil {
-		if !(k8serrors.IsNotFound(err) && !c.noCreate) {
-			return err
-		}
-
-		if !c.noCreate && len(c.size) == 0 {
-			return fmt.Errorf("when creating a resource, the size must be specified")
-		}
-
-		var obj metav1.Object
-
-		if c.createPVC {
-			obj, err = c.createUploadPVC()
-			if err != nil {
-				return err
-			}
-		} else {
-			obj, err = c.createUploadDataVolume()
-			if err != nil {
-				return err
-			}
-		}
-
-		c.cmd.Printf("%s %s/%s created\n", reflect.TypeOf(obj).Elem().Name(), obj.GetNamespace(), obj.GetName())
-	} else {
-		pvc, err = c.ensurePVCSupportsUpload(pvc)
-		if err != nil {
-			return err
-		}
-
-		c.cmd.Printf("Using existing PVC %s/%s\n", c.namespace, pvc.Name)
-	}
-
-	if c.createPVC {
-		if err := c.waitUploadServerReady(); err != nil {
-			return err
-		}
-	} else {
-		if err := c.waitDvUploadScheduled(); err != nil {
-			return err
-		}
-	}
-	if c.uploadProxyURL == "" {
-		c.uploadProxyURL, err = c.getUploadProxyURL()
-		if err != nil {
-			return err
-		}
-		if c.uploadProxyURL == "" {
-			return fmt.Errorf("uploadproxy URL not found")
-		}
-	}
-
-	u, err := url.Parse(c.uploadProxyURL)
-	if err != nil {
-		return err
-	}
-
-	if u.Scheme == "" {
-		c.uploadProxyURL = fmt.Sprintf("https://%s", c.uploadProxyURL)
-	}
-
-	c.cmd.Printf("Uploading data to %s\n", c.uploadProxyURL)
-
-	token, err := c.getUploadToken()
-	if err != nil {
-		return err
-	}
-
-	if err := c.uploadData(token, file); err != nil {
-		return err
-	}
-
-	if c.dataSource {
-		if err := c.handleDataSource(); err != nil {
-			return err
-		}
-	}
-
-	c.cmd.Println("Uploading data completed successfully, waiting for processing to complete, you can hit ctrl-c without interrupting the progress")
-	err = UploadProcessingCompleteFunc(c.client, c.cmd, c.namespace, c.name, processingWaitInterval, processingWaitTotal)
-	if err != nil {
-		c.cmd.Printf("Timed out waiting for post upload processing to complete, please check upload pod status for progress\n")
-	} else {
-		c.cmd.Printf("Uploading %s completed successfully\n", c.imagePath)
-	}
-
-	return err
 }
 
 func GetHTTPClient(insecure bool) *http.Client {
